layer2/requirements.txt
fastapi
uvicorn
pyTigerGraph
openai
python-dotenv
python-multipart
pandas
numpy
scikit-learn
requests